\name{patents}
\docType{data}
\alias{patents}
\title{Dataset of US patents from 1980 to 1999, citations distribution}
\description{
Dataset of US patents that start at 1980 to 1999 and have been cited at least 20 times, frequency distribution in years is given
}
\usage{patents}
\format{A dataframe of 3245 patents, each of 20 years frequency citation distribution}
\source{United State Patents and Trademark Office, 2005} 
\references{
Hall, B. H., Jaffe, A. B. and Tratjenberg, M. (2001) The NBER patent citation data file: Lessons, insights and methodological tools. NBER Working Paper 8498, 
NBER. \url{http://papers.nber.org/papers/w8498.pdf}.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datasets}
